requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.0.7
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
Pygments==2.7.4
PyJWT==1.7.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
wcwidth==0.2.5
